simul: handle psl endpoints (and adjust issue45).

+++ testsuite/gna/issue45/testsuite.sh
@@ -5,10 +5,14 @@
 GHDL_STD_FLAGS="--std=08 -fpsl"
 
 analyze test1.vhdl
-elab_simulate psl_test_endpoint --psl-report=psl.out
+elab psl_test_endpoint
 
-grep -q '"cover-pass": 3' psl.out
-rm psl.out
+if ghdl_has_feature psl_test_endpoint psl-report; then
+    simulate psl_test_endpoint --psl-report=psl.out
+
+    grep -q '"cover-pass": 3' psl.out
+    rm psl.out
+fi
 
 analyze test2.vhdl
 elab_simulate psl_test2_endpoint --assert-level=error --expect-failure
